use latest audits and scan info from data repo
